requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow-cpu
jupyter
flask
fastapi
uvicorn[standard]
jinja2
gunicorn
joblib
mlflow
